requirements/local.txt
asgiref~=3.5.2
Django~=4.0.6
django-model-utils~=4.2.0
djangorestframework~= 3.13.1
Pillow~=9.1.1
psycopg2~=2.9.3
tzdata~=2022.1
sqlparse~=0.4.2
Unipath~=1.1